Avoid division by zero in HTML templates when there are no uploads
For every country for every competition, the tool computes
the file usage rate and the percentage of new users.

This is computed in the Jinja templates by dividing by the
upload count and the uploader count, which results in
a ZeroDivisionError when there have been no uploads yet.

Now, if the upload[er] count is zero, we do not compute
and return a percentage of zero directly.