Default user_registration to 2005 if NULL
The query looks up the registratio date of the uploader.
But old-enough users on Wikimedia projects have no
registration date in the database.

It /probably/ does not matter too much for this tool,
as we are only interested in that to check whether
uploaders registered before or after the competition start

Just to be on the safe side, let's default default
the timestamp to 2005 for these users.