Replace database connection library MysqlDB with PyMySQL
`oursql` is a Python interface to MySQL that had not
been udpated in two years. It also needs MySql installed
locally, typically resulting when errors trying to run
locally (which is a pain when writing tests).

`PyMySQL` is a pure-Python MySQL client library,
which is an almost drop-in replacement to `oursql`.

See https://github.com/PyMySQL/PyMySQL