Do not reuse `data` structure in `getData`
The competition data structure (`data`) was being enriched
by udapting it, which prevents isolating behaviours and testing.

This adds a new data structure, `country_data` which is being
populated by the data processing.

For some reason, this caused the test to break because the result
now uses `defaultdict`. Bit unclear why it was not there before
but it appears to be what is intended.