Add logging of discarded images as out of bounds
The tool does not count pictures in a given category
that were uploaded outside of the competition dates.

When there in some mismatch between the Commons category
count and the tool count, it is usually thought to be
because of that, but could also be an actual bug.

This logging should help us understand what the images
considered out-of-bounds actually account for.