Align build/deploy script with Toolforge conventions
The Toolforge Python environment has conventions
on the layout of a Python project.

We were already mostly following them, but this
fully and explicitly complies with them.

- Add `TOOLFORGE_PATH` variable, pointing to `$HOME/www/python/`
- Change virtual environment path to match the convention
- Symlink the expected app path to our clone path during build

We could clone directly in ~/www/path/, but in the general
case there is no reason that the app.py would be at the top-level.

https://wikitech.wikimedia.org/wiki/Help:Toolforge/My_first_Flask_OAuth_tool