Allow to update database for a given list of events
While a contest is running, we want to update its data
as often as possible. However, updating all events
(including some finished for years) is fairly expensive
and cannot be done as it overloads the database servers.

This allows to pass a given list of events to update as
command-line arguments and only update those.

(The tool already knows how to incrementally update each dataset.)

Also switch to `argparse` for sane argument handling.

Bug: T203702