Enrich daily data with joiner count
An helpful indicator on the health of a competition,
more than the images count, is the uploader count.

To monitor during the competition, it would be helpful
to see per day how many users join the competition.

This enriches the daily data structure to hold for each
day, on top of the upload count, the joiner count, and
how many of them are newly registered users.

Bug: T203816