Ensure country names are normalized from the URL
Several URL routes match a country name, eg
</monuments/2018/United_Arab_Emirates/users>.

This country name is used to look up the right
data from the database (as dictionary keys).

When a country is composed of several words,
the URL may have underscores, resulting in the
country not being found.

This ensures that the underscores are replaced
with spaces first thing after reading from the URL.