Ensure country names are normalized from the URL in images.py
The country name is used to constuct the Commons category
to query. For some countries, we maintain a list of
special exceptions (see 907ecbcb).

When a country is composed of several words,
the URL may have underscores, resulting in the
country not being found in the special exceptions

This ensures that the country name is normalized
before the call to `get_wikiloves_category_name`.

Bug: T204427