Add Wiki Science Competition
The Wiki Science Competition is a federated photo
contest in the style of Wiki Loves X events.

It was not trivial adding it to the tool however,
as it is breaking some of the earlier assumptions.

Over the last year we have been able to generalize the
tool significantly. This allow us to easily add new scopes
and to accomodate the Wiki Science specificities.

- Add URL route and logo template
- Add event exception for `Wiki Science Competition (and tests)
- Add edition exception for `European Science Competition 2015` (and tests)
- Add special exception for `Images from Teadusfoto 201X`
- Update test fixtures and unit tests

Bug: T182956